cross tabs navigation
Improvements to navigate between tabs so as to understand numbers, or concepts.
From Income or Balance concepts by clicking on the tables, it goes to ledger showing what items sum up that concept.
From Ledger, by clicking on entry, it goes to journal showing the complete transaction.
From Journal, you can expand an account on Ledger.
This navigation happens always with the year shown in the origin tab.